Updated folder structure. Added setup.py. Added more to util.py, namely a more robust curry function.
