+++ .gitignore
@@ -1,0 +1,5 @@
+build
+dist
+*.egg*
+*.pyc
+__pycache__
